Extract database APIs related to closed CFDs into submodule
